requirements.txt
asgiref==3.8.1
Django==4.2.25
django-crispy-forms==2.4
djangorestframework==3.16.1
et_xmlfile==2.0.0
geographiclib==2.1
geopy==2.4.1
openpyxl==3.1.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
